model_tester_streamlit_app/requirements.txt
streamlit==1.10.0
openai==0.11.0
python-dotenv==0.19.2
pandas==1.3.4